+++ src-tauri/src/core/commands/mod.rs
@@ -1,1 +1,2 @@
+pub mod space;
 pub mod window;
