+++ src-tauri/src/core/mod.rs
@@ -1,1 +1,2 @@
+pub mod commands;
 pub mod space;
